header corregit
S'ha corregit el error que ens sortia amb el header, que no ens sortia el text roig quan pasavam per sobre el ratoli
