+++ Taxis/src/app/app-routing.module.ts
@@ -6,7 +6,6 @@ import { Error404Component } from './components/error404/error404.component';
 import { IniciComponent } from './components/inici/inici.component';
 import { QuiSomComponent } from './components/qui-som/qui-som.component';
 import { TrabomComponent } from './components/trobam/trabom.component';
-import { LoginComponent } from './components/ingres/login/login.component';
 import { SigninComponent } from './components/signin/signin.component';
 import { IngresComponent } from './components/ingres/ingres.component';
 
@@ -18,7 +17,6 @@ const routes: Routes = [
   { path: 'quisom', component: QuiSomComponent},
   { path: 'contacte', component: ContacteComponent},
   { path: 'companyia', component: CompanyiaComponent},
-  { path: 'login', component: LoginComponent},
   { path: 'signin', component: SigninComponent},
   { path: 'ingres', component: IngresComponent},
   { path: '**', component: Error404Component }
