+++ Taxis/src/app/app.module.ts
@@ -11,7 +11,6 @@ import { ContacteComponent } from './components/contacte/contacte.component';
 import { CompanyiaComponent } from './components/companyia/companyia.component';
 import { TrabomComponent } from './components/trobam/trabom.component';
 import { Error404Component } from './components/error404/error404.component';
-import { LoginComponent } from './components/ingres/login/login.component';
 import { SigninComponent } from './components/signin/signin.component';
 import { IngresComponent } from './components/ingres/ingres.component';
 
@@ -26,7 +25,6 @@ import { IngresComponent } from './components/ingres/ingres.component';
     CompanyiaComponent,
     TrabomComponent,
     Error404Component,
-    LoginComponent,
     SigninComponent,
     IngresComponent
   ],
