+++ Taxis/src/app/components/header/header.component.html
@@ -23,15 +23,15 @@
       </ul>
       <div class="form-inline my-2 my-lg-0">
         <ul class="navbar-nav mr-auto mt-2 mt-lg-0">
-            <li class="nav-item active">
-              <a class="nav-link" ng-if="conected" [routerLink]="['/ingres']">Log in</a>
-            </li>
-            <li class="nav-item active">
-                <a class="nav-link" ng-if="conected" [routerLink]="['/signin']">Registrar-se</a>
-            </li>
-            <li class="nav-item active">
-                <a class="nav-link" ng-if="disconected = false">Log out</a>
-              </li>
+          <li class="nav-item active">
+            <a class="nav-link" ng-if="conected" [routerLink]="['/ingres']">Log in</a>
+          </li>
+          <li class="nav-item active">
+              <a class="nav-link" ng-if="conected" [routerLink]="['/signin']">Registrar-se</a>
+          </li>
+          <li class="nav-item active">
+              <a class="nav-link" ng-if="disconected = false">Log out</a>
+          </li>
         </ul>
       </div>
     </div>
