+++ Taxis/src/app/components/ingres/ingres.component.css
@@ -1,9 +1,9 @@
 .all {
-    background-image: url(/assets/general/background.png);
+    /* background-image: url(/assets/general/background.png); */
     height: 100%;
     width: 100%;
-    position: fixed;
-    top: 0;
+    position: relative;
+    transform: translateY(100%);
 }
 
 .center {
