+++ Taxis/src/app/components/ingres/ingres.component.ts
@@ -7,9 +7,30 @@ import { Component, OnInit } from '@angular/core';
 })
 export class IngresComponent implements OnInit {
 
+  verpasswd: boolean = false;
+
   constructor() { }
 
   ngOnInit(): void {
+    
+  }
+
+  habilitarpasswd(): void {
+
+    let campo:any = document.getElementById('passwd');
+
+    if (this.verpasswd) {
+
+      this.verpasswd = false;
+      campo.type = "password";
+
+    } else {
+
+      this.verpasswd = true;
+      campo.type = "text";
+
+    }
+
   }
 
 }
