--- Taxis/src/app/components/ingres/login/login.component.html
@@ -1,42 +1,0 @@
-<!-- Formulario de Log In -->
-
-<form class="container">
-
-    <div class="form-group">
-
-      <input type="text" class="form-control" placeholder="Nickname">
-
-        <!-- Validar Nickname
-        <div *ngIf = "submitted && validacion.nickname.errors" class="invalid-feedback mensaje-error">
-
-          <p *ngIf = "validacion.nickname.errors.required">Este campo es obligatorio</p>
-
-        </div> -->
-
-    </div>
-
-    <div class="form-group">
-
-      <div class="input-group mb-2">
-
-        <input id="passwd" type="password" class="form-control" placeholder="Contraseña">
-
-        <div class="input-group-prepend">
-
-          <div class="input-group-text btn" (click)="habilitarpasswd()">
-
-            <img *ngIf="verpasswd == false" src="/assets/icons/eye.png" width="18px" title="Ver contraseña">
-
-            <img *ngIf="verpasswd == true" src="/assets/icons/eye-off.png" width="18px" title="Ocultar contraseña">
-
-          </div>
-
-        </div>
-
-      </div>
-
-    </div>
-
-  <button type="submit" class="btn btn-outline-info btn-block">Entrar</button>
-
-</form>
